initialize

+++ py/DeepFunction.py
@@ -17,7 +17,7 @@ class DevBuffer:
 	def toHost(self, is_blocking = True):
 		if self.devTT.pos != "h":
 			self.devTT.pos = "h"
-			self.devTT.host_buf = np.empty(self.devTT.shape, dtype = np.float16)
+			self.devTT.host_buf = np.empty(self.devTT.shape, dtype = np.float32)
 			pyCL.enqueue_copy(self.devTT.deep_func.q, self.devTT.host_buf, self.devTT.buf, is_blocking = is_blocking)
 		return self.devTT.host_buf
